+++ nix/modules/flake-module.nix
@@ -6,12 +6,13 @@
   ...
 }:
 let
-  inherit (flake-parts-lib) mkPerSystemOption;
-  inherit (lib) mkOption types;
-  defaultSpecialArgs = (import ../lib/lib.nix { inherit inputs; }).conanFlake.defaultSpecialArgs {
-    inherit lib;
-    infuse = (import inputs.infuse { inherit lib; }).v1.infuse;
-  };
+  inherit (flake-parts-lib)
+    mkPerSystemOption
+    ;
+  inherit (lib)
+    mkOption
+    ;
+  conan-flake-lib = (import ../lib/lib.nix { inherit inputs; }).conanFlakeLib;
 in
 {
   options = {
@@ -26,48 +27,29 @@ in
           description = ''
             Project-level Conan configuration
 
-            Use `config.treefmt.build.wrapper` to get access to the resulting treefmt
+            Use `config.conan.build.wrapper` to get access to the resulting Conan
             package based on this configuration.
-
-            By default treefmt-nix will set the `formatter.<system>` attribute of the flake,
-            used by the `nix fmt` command.
           '';
-          type = (
-            types.submoduleWith {
-              specialArgs = {
-                inherit pkgs;
-                inherit (defaultSpecialArgs)
-                  infuse
-                  relativePathType
-                  parseSystemArch
-                  parseSystemOs
-                  envSubmodule
-                  listOfGeneratorType
-                  outputType
-                  listOfOutputType
-                  anyOutput
-                  ;
-              };
-              modules = [
-                ./configuration
-                {
-                  configRoot = lib.mkDefault self;
-                }
-              ];
-            }
-          );
+          type = conan-flake-lib.submoduleWith lib {
+            modules = [
+              {
+                options.pkgs = lib.mkOption {
+                  default = pkgs;
+                  defaultText = lib.literalMD "`pkgs` (module argument of `perSystem`)";
+                };
+
+                config.configRoot = lib.mkDefault self;
+              }
+            ];
+          };
           default = { };
         };
-        config =
-          let
-            contains = k: lib.any (x: x == k);
-          in
-          {
-            devShells = lib.optionalAttrs (contains "devShells" config.conan.autoWire) {
-              configuration = config.conan.outputs.devShell;
-            };
-            checks = lib.optionalAttrs (contains "checks" config.conan.autoWire) config.conan.outputs.checks;
+        config = {
+          devShells = lib.optionalAttrs (conan-flake-lib.contains "devShells" config.conan.autoWire) {
+            configuration = config.conan.outputs.devShell;
           };
+          checks = lib.optionalAttrs (conan-flake-lib.contains "checks" config.conan.autoWire) config.conan.outputs.checks;
+        };
       }
     );
   };